Tested create/write/delete of files. Added 'touch' and 'rm' telnet commands. Modified the SD initialisation a little.

+++ web-platform/firmware/uIPStack/lib/dosfs.h
@@ -11,8 +11,8 @@
 
 //===================================================================
 // User-supplied functions
-uint32_t DFS_ReadSector(uint8_t unit, uint8_t *buffer, uint32_t sector, uint32_t count);
-uint32_t DFS_WriteSector(uint8_t unit, uint8_t *buffer, uint32_t sector, uint32_t count);
+uint8_t DFS_ReadSector(uint8_t unit, uint8_t *buffer, uint32_t sector, uint32_t count);
+uint8_t DFS_WriteSector(uint8_t unit, uint8_t *buffer, uint32_t sector, uint32_t count);
 
 
 //===================================================================
